+++ src/metadata.php
@@ -22,12 +22,13 @@ $aModule = [
     'version'     => '4.4.1',
     'thumbnail'   => 'out/pictures/logo.png',
     'controllers' => [
-        'ffWebComponent'     => Component\Widget\WebComponent::class,
-        'search_result'      => Controller\SearchResultController::class,
-        'test_connection'    => Controller\Admin\TestConnectionController::class,
-        'article_feed'       => Controller\Admin\ArticleFeedController::class,
-        'http_article_feed'  => Controller\ArticleFeedController::class,
-        'http_category_feed' => Controller\CategoryFeedController::class,
+        'ffWebComponent'             => Component\Widget\WebComponent::class,
+        'search_result'              => Controller\SearchResultController::class,
+        'test_connection'            => Controller\Admin\TestConnectionController::class,
+        'article_feed'               => Controller\Admin\ArticleFeedController::class,
+        'http_article_feed'          => Controller\ArticleFeedController::class,
+        'http_category_feed'         => Controller\CategoryFeedController::class,
+        'http_suggest_category_feed' => Controller\SuggestCategoryFeedController::class,
     ],
     'blocks'      => [
         [