+++ tests/Unit/Export/Exporter/ExportEntitiesTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\Export\Exporter;
+namespace FactFinderTests\Unit\Export\Exporter;
 
 use Omikron\FactFinder\Oxid\Export\Data\ArticleCollection;
 use Omikron\FactFinder\Oxid\Export\Entity\DataProvider;
@@ -11,9 +11,9 @@ use OxidEsales\Eshop\Application\Model\Article;
 use OxidEsales\Eshop\Core\Field;
 use OxidEsales\Eshop\Core\Registry;
 use PHPUnit\Framework\TestCase;
-use Tests\Variant\Export\Data\ArticleCollectionVariant;
-use Tests\Variant\Export\Field\DisplayError;
-use Tests\Variant\Export\Stream\CsvVariant;
+use FactFinderTests\Variant\Export\Data\ArticleCollectionVariant;
+use FactFinderTests\Variant\Export\Field\DisplayError;
+use FactFinderTests\Variant\Export\Stream\CsvVariant;
 
 class ExportEntitiesTest extends TestCase
 {
@@ -59,6 +59,11 @@ class ExportEntitiesTest extends TestCase
             'oxarticles__oxshortdesc' => new Field('bicycle short description'),
             'oxarticles__oxartnum' => new Field('1500'),
             'oxarticles__oxprice' => new Field('2000.00'),
+            'oxarticles__oxparentid' => new Field(''),
+            'oxarticles__oxvarselect' => new Field(''),
+            'oxarticles__oxpic1' => new Field('bicycle_1.jpg'),
+            'oxarticles__oxskipdiscounts' => new Field('0'),
+            'oxarticles__oxvarcount' => new Field('10'),
         ]);
         $products = new ArticleCollectionVariant([$article]);
         $dataProvider = new DataProvider($products);
@@ -81,6 +86,11 @@ class ExportEntitiesTest extends TestCase
             'oxarticles__oxshortdesc' => new Field('bicycle short description'),
             'oxarticles__oxartnum' => new Field('1500'),
             'oxarticles__oxprice' => new Field('2000.00'),
+            'oxarticles__oxparentid' => new Field(''),
+            'oxarticles__oxvarselect' => new Field(''),
+            'oxarticles__oxpic1' => new Field('bicycle_1.jpg'),
+            'oxarticles__oxskipdiscounts' => new Field('0'),
+            'oxarticles__oxvarcount' => new Field('10'),
         ]);
         $articleTwo = new Article([
             'oxarticles__oxtitle' => new Field('skateboard'),
@@ -88,6 +98,11 @@ class ExportEntitiesTest extends TestCase
             'oxarticles__oxshortdesc' => new Field('skateboard short description'),
             'oxarticles__oxartnum' => new Field('1501'),
             'oxarticles__oxprice' => new Field('300.00'),
+            'oxarticles__oxparentid' => new Field(''),
+            'oxarticles__oxvarselect' => new Field(''),
+            'oxarticles__oxpic1' => new Field('skateboard_1.jpg'),
+            'oxarticles__oxskipdiscounts' => new Field('0'),
+            'oxarticles__oxvarcount' => new Field('10'),
         ]);
         $products = new ArticleCollectionVariant([$articleOne, $articleTwo]);
         $dataProvider = new DataProvider($products);
@@ -118,6 +133,11 @@ class ExportEntitiesTest extends TestCase
             'oxarticles__oxartnum' => new Field('1500'),
             'oxarticles__oxprice' => new Field('2000.00'),
             'oxarticles__oxdisplayerror' => new Field('Sample exception message'),
+            'oxarticles__oxparentid' => new Field(''),
+            'oxarticles__oxvarselect' => new Field(''),
+            'oxarticles__oxpic1' => new Field('bicycle_1.jpg'),
+            'oxarticles__oxskipdiscounts' => new Field('0'),
+            'oxarticles__oxvarcount' => new Field('10'),
         ]);
         $articleTwo = new Article([
             'oxarticles__oxtitle' => new Field('skateboard'),
@@ -125,6 +145,11 @@ class ExportEntitiesTest extends TestCase
             'oxarticles__oxshortdesc' => new Field('skateboard short description'),
             'oxarticles__oxartnum' => new Field('1501'),
             'oxarticles__oxprice' => new Field('300.00'),
+            'oxarticles__oxparentid' => new Field(''),
+            'oxarticles__oxvarselect' => new Field(''),
+            'oxarticles__oxpic1' => new Field('skateboard_1.jpg'),
+            'oxarticles__oxskipdiscounts' => new Field('0'),
+            'oxarticles__oxvarcount' => new Field('10'),
         ]);
         $products = new ArticleCollectionVariant([$articleOne, $articleTwo]);
         $dataProvider = new DataProvider($products, new DisplayError());
@@ -139,10 +164,6 @@ class ExportEntitiesTest extends TestCase
         // Expect & Given
         $logFilename = sprintf('%s/log/fact-finder/exporter.log', SHOP_SOURCE_PATH);
 
-        if (file_exists($logFilename)) {
-            unlink($logFilename);
-        }
-
         $columns = array_merge($this->columns, ['DisplayError']);
         $stream = new CsvVariant($this->tmpfile);
         $articleOne = new Article([
@@ -151,6 +172,11 @@ class ExportEntitiesTest extends TestCase
             'oxarticles__oxshortdesc' => new Field('bicycle short description'),
             'oxarticles__oxartnum' => new Field('1500'),
             'oxarticles__oxprice' => new Field('2000.00'),
+            'oxarticles__oxparentid' => new Field(''),
+            'oxarticles__oxvarselect' => new Field(''),
+            'oxarticles__oxpic1' => new Field('bicycle_1.jpg'),
+            'oxarticles__oxskipdiscounts' => new Field('0'),
+            'oxarticles__oxvarcount' => new Field('10'),
         ]);
         $articleTwo = new Article([
             'oxarticles__oxtitle' => new Field('skateboard'),
@@ -159,6 +185,11 @@ class ExportEntitiesTest extends TestCase
             'oxarticles__oxartnum' => new Field('1501'),
             'oxarticles__oxprice' => new Field('300.00'),
             'oxarticles__oxdisplayerror' => new Field('Sample exception message'),
+            'oxarticles__oxparentid' => new Field(''),
+            'oxarticles__oxvarselect' => new Field(''),
+            'oxarticles__oxpic1' => new Field('skateboard_1.jpg'),
+            'oxarticles__oxskipdiscounts' => new Field('0'),
+            'oxarticles__oxvarcount' => new Field('10'),
         ]);
         $articleThree = new Article([
             'oxarticles__oxtitle' => new Field('stand up paddle'),
@@ -166,6 +197,11 @@ class ExportEntitiesTest extends TestCase
             'oxarticles__oxshortdesc' => new Field('stand up paddle short description'),
             'oxarticles__oxartnum' => new Field('1502'),
             'oxarticles__oxprice' => new Field('900.00'),
+            'oxarticles__oxparentid' => new Field(''),
+            'oxarticles__oxvarselect' => new Field(''),
+            'oxarticles__oxpic1' => new Field('stand-up-paddle_1.jpg'),
+            'oxarticles__oxskipdiscounts' => new Field('0'),
+            'oxarticles__oxvarcount' => new Field('10'),
         ]);
         $products = new ArticleCollectionVariant([$articleOne, $articleTwo, $articleThree]);
         $dataProvider = new DataProvider($products, new DisplayError());
@@ -199,6 +235,11 @@ class ExportEntitiesTest extends TestCase
             'oxarticles__oxshortdesc' => new Field('bicycle short description'),
             'oxarticles__oxartnum' => new Field('1500'),
             'oxarticles__oxprice' => new Field('2000.00'),
+            'oxarticles__oxparentid' => new Field(''),
+            'oxarticles__oxvarselect' => new Field(''),
+            'oxarticles__oxpic1' => new Field('bicycle_1.jpg'),
+            'oxarticles__oxskipdiscounts' => new Field('0'),
+            'oxarticles__oxvarcount' => new Field('10'),
         ]);
         $articleTwo = new Article([
             'oxarticles__oxtitle' => new Field('skateboard'),
@@ -207,6 +248,11 @@ class ExportEntitiesTest extends TestCase
             'oxarticles__oxartnum' => new Field('1501'),
             'oxarticles__oxprice' => new Field('300.00'),
             'oxarticles__oxdisplayerror' => new Field('Sample exception message'),
+            'oxarticles__oxparentid' => new Field(''),
+            'oxarticles__oxvarselect' => new Field(''),
+            'oxarticles__oxpic1' => new Field('skateboard_1.jpg'),
+            'oxarticles__oxskipdiscounts' => new Field('0'),
+            'oxarticles__oxvarcount' => new Field('10'),
         ]);
         $articleThree = new Article([
             'oxarticles__oxtitle' => new Field('stand up paddle'),
@@ -214,6 +260,11 @@ class ExportEntitiesTest extends TestCase
             'oxarticles__oxshortdesc' => new Field('stand up paddle short description'),
             'oxarticles__oxartnum' => new Field('1502'),
             'oxarticles__oxprice' => new Field('900.00'),
+            'oxarticles__oxparentid' => new Field(''),
+            'oxarticles__oxvarselect' => new Field(''),
+            'oxarticles__oxpic1' => new Field('stand-up-paddle_1.jpg'),
+            'oxarticles__oxskipdiscounts' => new Field('0'),
+            'oxarticles__oxvarcount' => new Field('10'),
         ]);
         $products = new ArticleCollectionVariant([$articleOne, $articleTwo, $articleThree]);
         $dataProvider = new DataProvider($products, new DisplayError());