+++ tests/Unit/Subscriber/AfterRequestProcessedEventSubscriber/hasJustLoggedInTest.php
@@ -2,10 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\Subscriber\AfterRequestProcessedEventSubscriber;
+namespace FactFinderTests\Unit\Subscriber\AfterRequestProcessedEventSubscriber;
 
 use Omikron\FactFinder\Oxid\Subscriber\AfterRequestProcessedEventSubscriber;
+use Omikron\FactFinder\Oxid\Subscriber\BeforeHeadersSendEventSubscriber;
 use OxidEsales\Eshop\Core\Config;
+use OxidEsales\Eshop\Core\Request;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use OxidEsales\Eshop\Core\Session;
@@ -13,10 +15,13 @@ use OxidEsales\Eshop\Application\Model\User;
 
 class hasJustLoggedInTest extends TestCase
 {
-    /** @var MockObject  */
+    /** @var Request|MockObject  */
+    private $request;
+
+    /** @var Session|MockObject  */
     private $session;
 
-    /** @var MockObject  */
+    /** @var Config|MockObject  */
     private $config;
 
     public function testShouldSetSessionVariableWhenUserIsSetAndActionLoginNoRedirect()
@@ -24,7 +29,7 @@ class hasJustLoggedInTest extends TestCase
         // Expect
         $this->config->method('getRequestParameter')->with('fnc')->willReturn('login_noredirect');
         $this->session->method('getUser')->willReturn($this->createMock(User::class));
-        $this->session->expects($this->once())->method('setVariable')->with('ff_has_just_logged_in', true);
+        $this->session->expects($this->once())->method('setVariable')->with(BeforeHeadersSendEventSubscriber::HAS_JUST_LOGGED_IN, true);
 
         // When & Then
         $this->subscriber->hasJustLoggedIn();
@@ -35,7 +40,7 @@ class hasJustLoggedInTest extends TestCase
         // Expect
         $this->config->method('getRequestParameter')->with('fnc')->willReturn('login_noredirect');
         $this->session->method('getUser')->willReturn(false);
-        $this->session->expects($this->never())->method('setVariable')->with('ff_has_just_logged_out', true);
+        $this->session->expects($this->never())->method('setVariable')->with(BeforeHeadersSendEventSubscriber::HAS_JUST_LOGGED_OUT, true);
 
         // When & Then
         $this->subscriber->hasJustLoggedIn();
@@ -46,7 +51,7 @@ class hasJustLoggedInTest extends TestCase
         // Expect
         $this->config->method('getRequestParameter')->with('fnc')->willReturn('some_action');
         $this->session->method('getUser')->willReturn($this->createMock(User::class));
-        $this->session->expects($this->never())->method('setVariable')->with('ff_has_just_logged_out', true);
+        $this->session->expects($this->once())->method('setVariable')->with(BeforeHeadersSendEventSubscriber::HAS_JUST_LOGGED_IN, true);
 
         // When & Then
         $this->subscriber->hasJustLoggedIn();
@@ -54,13 +59,13 @@ class hasJustLoggedInTest extends TestCase
 
     protected function setUp(): void
     {
-        $this->session = $this->getMockBuilder(Session::class)
-            ->disableOriginalConstructor()
-            ->onlyMethods(['setVariable', 'getUser'])
-            ->getMock();
-        $this->config = $this->getMockBuilder(Config::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-        $this->subscriber = new AfterRequestProcessedEventSubscriber($this->session, $this->config);
+        $this->request = $this->createMock(Request::class);
+        $this->session = $this->createMock(Session::class);
+        $this->config = $this->createMock(Config::class);
+        $this->subscriber = new AfterRequestProcessedEventSubscriber(
+            $this->request,
+            $this->session,
+            $this->config
+        );
     }
 }