+++ tests/Variant/Export/Field/DisplayError.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Tests\Variant\Export\Field;
+namespace FactFinderTests\Variant\Export\Field;
 
-use Omikron\FactFinder\Oxid\Export\Field\FieldInterface;
+use Omikron\FactFinder\Oxid\Export\Field\Article\FieldInterface;
 use OxidEsales\Eshop\Application\Model\Article;
 
 class DisplayError implements FieldInterface